% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{contains_covariates}
\alias{contains_covariates}
\title{Determine if a given parameter uses covariates}
\usage{
contains_covariates(base, par, DMs)
}
\arguments{
\item{base}{The character name of the type of parameter. "mu", "sigma", or
"w"}

\item{par}{A vector of parameter estimates, typically supplied by optim}

\item{DMs}{The list of design matrices which contains all of the design
matrices required to calculated the value of each parameter given its
covariate values}
}
\value{
A boolean. TRUE if the parameter "base" uses covariates, and 
FALSE otherwise.
}
\description{
Parses through a GAI model's skeleton and design matrix list to determine
 which type of covariate specification has been used (No covariates, the
 same formula for each brood, or brood-specific formulas)
}
