% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{dzip}
\alias{dzip}
\title{Zero-Inflated Poisson Density Evaluation}
\usage{
dzip(obs, lambda, psi, log = T)
}
\arguments{
\item{obs}{The observations. Can be a matrix or vector, and may contain NAs}

\item{lambda}{The vector of means for each observation. Must have the same 
length as obs}

\item{psi}{The probability of being from the poisson distribution as
opposed to being an 'inflated' 0}

\item{log}{A boolean, if TRUE the function returns the logs of the
densities instead}
}
\value{
A vector (or matrix) of the same length (dimension) as obs, with
the density or log density of the zero-inflated poisson evaluated
at each point.
}
\description{
Calculates the density of a zero-inflated poisson model at given data points
for a set of parameter values
}
