% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{evaluate_splines}
\alias{evaluate_splines}
\title{Spline model density evaluation.}
\usage{
evaluate_splines(par, skeleton, spline_specs, nT, nS)
}
\arguments{
\item{par}{The vector of estimated spline coefficients (and also potentially
the distributional parameters for ZIP and NB likelihoods)}

\item{skeleton}{The skeleton to \code{\link{relist}} the parameters}

\item{spline_specs}{The list containing the 'degree' of the splines and 
also df (their degrees of freedom)}

\item{nT}{The number of sampling occasions per site}

\item{nS}{The number of sampling locations during the survey}
}
\value{
A vector of values of length nT, giving the relative density
of expected counts at each occasion
}
\description{
Calculates the seasonal flight curve using splines
}
