% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{evaluate_stopover}
\alias{evaluate_stopover}
\title{Stopover model density evaluation.}
\usage{
evaluate_stopover(breaks, par, skeleton, DMs = list(), nS)
}
\arguments{
\item{breaks}{The points to evaluate. The function return the probability of
being in each bin defined by the breaks.}

\item{par}{A vector of parameter values which should be
\code{\link{relist}}ed}

\item{skeleton}{The skeleton to \code{\link{relist}} the parameters}

\item{DMs}{A list of design matrices to be used when covaraites are included
in the model.}

\item{nS}{In the case where DMs is empty because we use no covariates, we
need to know the number of sites to determine the number of points.}
}
\value{
A matrix of evaluated densities for the count at each site and
location, or the single real number valued negative log likelihood
}
\description{
Evaluates the CDF of a mixture of normal distributions, then
returns the probability of being in a given bin.
}
