% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{extract_counts}
\alias{extract_counts}
\title{Extract a count matrix from a data.frame}
\usage{
extract_counts(data_frame, checks = T, returnDF = T)
}
\arguments{
\item{data_frame}{A data.frame with columns "site", "occasion" and "counts"}

\item{checks}{if TRUE, checks that all counts are specified}

\item{returnDF}{if TRUE returns instead a list with the matrix of counts as
well as a reordered version of the data.frame so that the rows match the
matrix output.}
}
\value{
A matrix of counts, with rows as sites and columns as occasions.
}
\description{
Loops through unique values of 'occasion' and 'site' in a data.frame in order
to extract the matrix of observed counts required for GAI model fitting
}
\examples{
extract_counts(example_data)
}
