% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{fit_GAI}
\alias{fit_GAI}
\title{GAI model fitting}
\usage{
fit_GAI(
  start,
  DF,
  a_choice = "mixture",
  dist_choice = "P",
  options = list(),
  tol = 0.001,
  maxiter = 1000,
  verbose = F,
  hessian = F,
  bootstrap = NULL
)
}
\arguments{
\item{start}{The vector of start guesses for the parameters.}

\item{DF}{The data.frame of observations. Should contain columns 'site', 
'occasion' and 'count'. occasion and count should be integers, and missing 
entries in the count column are allowed to take the value NA. Any covariate
which is referred to in a formula specified in the options argument should
be present as a column in DF.}

\item{a_choice}{"mixture", "stopover" or "splines".}

\item{dist_choice}{"P" (poisson), "NB" (negative binomial) or "ZIP"
(zero-inflated poisson).}

\item{options}{A list containing different specifications, which vary
depending on the model. For stopover and mixture models this contains B (the
number of broods), shared_sigma (boolean denoting if the SDs are the same for
each component), mu_formula (specifying a formula which describes a covariate
dependency for the mean arrival times for each brood), and sd_formula
(similar, for the SD of each brood).}

\item{tol}{The tolerance for the NB and ZIP iterative solvers. The solver
stops when the difference between the negative log likelihood from one
iteration to the next falls beneath this threshold.}

\item{maxiter}{The maximum number of iterations to perform when iterating for
the NB or ZIP likelihoods}

\item{verbose}{If TRUE, the function will print various updates while it
iterates for NB and for ZIP models.}

\item{hessian}{if TRUE, refits the model one last time with hessian = TRUE in
optim, usually so that this can be used to conduct a bootstrap}

\item{bootstrap}{Always NULL for the user.. Internally, bootstraps use
this argument to pass in sanitised precaulcations to speed up the refitting
process of non-parametric bootstraps}
}
\value{
a fitted optim object
}
\description{
Fits the GAI model to a set of data, with a choice of normal, stopover or
spline fits for the seasonal arrival times, as well as poisson, zero-inflated
poisson, or negative binomially distribted count observations.
}
