% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\docType{package}
\name{GAI}
\alias{GAI}
\title{GAI: Generalised Abundance Index fitting for seasonal invertibrates, with
covariates}
\description{
Fitting of the GAI from Dennis et al (2016), with normal mixture, stopover
and spline-based seasonal trend models, as well as poisson, negative binomial
and zero-inflated count models. Covariates can be fitted to mean brood
arrival time, mean brood dispersion and the proportion of the population
which finds itself in each brood. The package slightly extends the
supplementary materials of the paper to provide flexible model-fitting
options and covariate inclusion.
}
\details{
The package should be used through the fit_GAI function.
Options should be specified to give more detail of the model being fitted.
Bootstraps can be used to produce confidence interval estimates of the 
abundance index and parameter values.
}
\author{
C. Fagard-Jenkin, based on work by E. Dennis, et al.
}
