% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{generate_mixture_pars}
\alias{generate_mixture_pars}
\title{Generate random parameter values for non-spline models}
\usage{
generate_mixture_pars(nT, skeleton, sd_max = 1, l = -5, u = 5)
}
\arguments{
\item{nT}{The number of sampling occasions in the experiment}

\item{skeleton}{The skeleton to \code{\link{relist}} the parameters}

\item{sd_max}{The maximum SD for broods}

\item{l}{The lower limit for the distributional parameter (before the link
function is applied)}

\item{u}{The upper limit for the distributional parameter (before the link
function is applied)}
}
\value{
A vector of parameters for the model, such that it can be relisted
using skeleton
}
\description{
For ease of simulation studies or testing, generates some parameters for 
non-covariate including normal mixture and stopover models
}
