% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{generate_spline_pars}
\alias{generate_spline_pars}
\title{Generate random parameter values for spline models}
\usage{
generate_spline_pars(skeleton, sigma = 2, l = -5, u = 5)
}
\arguments{
\item{skeleton}{The skeleton to \code{\link{relist}} the parameters}

\item{sigma}{The standard deviation of the normal distribution the parameters}

\item{l}{The lower limit for the dist.par parameter on the link scale}

\item{u}{The upper limit for the dist.par parameter on the link scale
are generated from}
}
\value{
A vector of parameters for the model, such that it can be relisted
using skeleton
}
\description{
For ease of simulation studies or testing, generates some parameters for 
non-covariate including spline models
}
