% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_cov_skeleton_and_DM}
\alias{get_cov_skeleton_and_DM}
\title{Obtain skeleton and Design Matrix}
\usage{
get_cov_skeleton_and_DM(base, options, DF, n, skeleton, DMs)
}
\arguments{
\item{base}{The character name of the type of parameter to be used. Can be
'mu', 'sigma' or 'w'.}

\item{options}{A list of options given to the fit_GAI function to 
specify the model}

\item{DF}{The data frame with which covariate values are stored, for creating
design matrices}

\item{n}{The number of formulas to check (i.e. how many parameters of type
'base' are we expecting?)}

\item{skeleton}{The object where skeleton items should be added}

\item{DMs}{The object where design matrix should be added}
}
\value{
A named list containing the updated skeleton and DMs objects in 
entries 'skeleton' and 'DMs' respectively.
}
\description{
Updates the skeleton and design matrix objects for a GAI model for a given 
input parameter type.
}
