% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_parameter_values}
\alias{get_parameter_values}
\title{Obtain parameter values and handle covariate formulas}
\usage{
get_parameter_values(par, DMs, skeleton, n)
}
\arguments{
\item{par}{A vector of parameter estimates, typically supplied by optim}

\item{DMs}{The list of design matrices which contains all of the design
matrices required to calculated the value of each parameter given its
covariate values}

\item{skeleton}{A named list of parameters which can be used to relist the 
vector of parameter guesses provided by optim into the structure expected by
GAI package functions}

\item{n}{The integer number of data points for which these parameter values
should be found}
}
\value{
A name dlist containing the Mu, Sigma and W parameters for the 
data points
}
\description{
Uses a suite of helper functions to calculate linear predictors for 
 parameter values, with and without covariate inclusion, for stopover and 
 mixture models
}
