% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_parameter_values_all_rows}
\alias{get_parameter_values_all_rows}
\title{Parameter value finding helper for covariate inclusion}
\usage{
get_parameter_values_all_rows(base, par, DMs, B)
}
\arguments{
\item{base}{The character name of the type of parameter. "mu", "sigma", or
"w"}

\item{par}{A vector of parameter estimates, typically supplied by optiim}

\item{DMs}{The list of design matrices which contains all of the design
matrices required to calculated the value of each parameter given its
covariate values}

\item{B}{The number of parameters for which the base needs to be used.
Typically the same as the number of broods, i.e. one mean arrival time per
brood,}
}
\value{
A matrix of pwrameter values to be used by get_parameter_values
}
\description{
A helper function which uses lp_func to get the values of parameters with 
 a covariate formula, for 
 all sites.
}
