% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{logLik.GAI}
\alias{logLik.GAI}
\title{Log likelihood extraction}
\usage{
\method{logLik}{GAI}(GAIobj)
}
\arguments{
\item{GAIobj}{An object produced by using <- for model fitting}
}
\value{
An object object of class "logLik" with an attribute "df" equal to 
the number of fitted parameters for the model at hand.
}
\description{
Extracts the (postive) log likelihood value from an output from the fit_GAI
function, notably for use by the AIC function
}
