% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{lp_func}
\alias{lp_func}
\title{Parameter value finding helper for covariate inclusion}
\usage{
lp_func(brood, is_indiv, is_general, par, base, DMs)
}
\arguments{
\item{brood}{the parameter number (integer) within the sequence. If being
used by the package internally, this is almost always the brood number}

\item{is_indiv}{A boolean indicating if the parameter for this given brood is
specified with a brood specific formula}

\item{is_general}{A boolean indicating if the parameter for this given is the
same as that of the other broods}

\item{par}{A vector of parameter estimates, typically supplied by optiim}

\item{base}{The character name of the type of parameter. "mu", "sigma", or
"w"}

\item{DMs}{The list of design matrices which contains all of the design
matrices required to calculated the value of each parameter given its
covariate values}
}
\value{
A matrix of parameter values, to be used by
get_parameter_values_all_rows
}
\description{
A helper function which obtains the linear predictor for a given set of
 parameters ('mu', 'sigma') etc, by scanning the list of design matrices and
 the model skeleton for the relevant information
}
