% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{means_backtransform}
\alias{means_backtransform}
\title{Backtransform for means}
\usage{
means_backtransform(vec)
}
\arguments{
\item{vec}{The vector of (positive) means which should be backtransformed to 
the real line}
}
\value{
A vector of the same length as vec, of backtransformed values
}
\description{
Performs the backtransform of the link function which is used to specify the
means of the broods
}
