% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{means_link}
\alias{means_link}
\title{Mean link function}
\usage{
means_link(m)
}
\arguments{
\item{m}{A vector of real valued mean parameters}
}
\value{
A vector ascending mean values, starting at m[1]
}
\description{
Transforms a vector of numbers on the real line to a vector of means that
have ascending values
}
