% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{mixture_nll}
\alias{mixture_nll}
\title{Normal mixture negative Log Likelihood}
\usage{
mixture_nll(par, obs, skeleton, returnNLL = T, DMs = list())
}
\arguments{
\item{par}{A list of parameters including mean, sd, and probs.
Alternatively, a vector of parameters, with a skeleton to relist the
parameters (parameter: skeleton).}

\item{obs}{A matrix of observed values.}

\item{skeleton}{The skeleton to \code{\link{relist}} the parameters}

\item{returnNLL}{If TRUE, returns the value of the NLL rather than the
likelihood for each point individually. If FALSE, also returns the output as
an nS*nT matrix.}

\item{DMs}{A named list of design matrices for covariate inclusion}
}
\value{
A matrix of evaluated densities for the count at each site and
location, or the single real number valued negative log likelihood
}
\description{
Evaluates an n-mixture model negative log-likelihood given the parameters and
the observations
}
