% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plot.GAI}
\alias{plot.GAI}
\title{GAI flight path plotting}
\usage{
\method{plot}{GAI}(
  GAIobj,
  all_sites = F,
  quantiles = c(0.05, 0.5, 0.95),
  scale_by_N = T,
  colours = 1:8,
  ...
)
}
\arguments{
\item{GAIobj}{An object produced by using the summary.GAI function an a model
output produce by fit_GAI.}

\item{all_sites}{If TRUE, the curves for every single site are plotted rather
than a smooth through the curves from each site. Defaults to FALSE.}

\item{quantiles}{For when all_sites = F, a vector of quantiles that will each
be used to produce a curve. Defaults to c(0.05, 0.5, 0.95)}

\item{scale_by_N}{If TRUE, then the seasonal flight component for each site
will be scaled by that site's estimated site total. Defaults to TRUE.}

\item{colours}{A vector of colour values to be used for plotting the
different curves. Defaults to integers 1:8 to represent the R base graphics
colours, but can also take character hex colour values.}
}
\description{
Prints the flight path curves of a fitted GAI model with options to scale
by estimated site total, and avoid plotting all sites by smoothing through
results for each sites by taking a given quantile at each time point
}
