% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{print.summary.GAI}
\alias{print.summary.GAI}
\title{GAI model summary printing}
\usage{
\method{print}{summary.GAI}(obj)
}
\arguments{
\item{obj}{An object produced by using the summary.GAI function an a model
output produce by fit_GAI.}
}
\value{
An object object of class "summary.GAI"
}
\description{
Prints the output of a GAI model summary, to display the information to the
end user.
}
