% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{probs_backtransform}
\alias{probs_backtransform}
\title{Backtransform for weights}
\usage{
probs_backtransform(vec, tol = 1e-10)
}
\arguments{
\item{vec}{The vector of weights between 0 and 1 which should be
backtransformed to the link-scale space}
}
\value{
A vector of the same length as vec, of backtransformed values
}
\description{
Performs the backtransform of the link function which is used to specify the
weights of the broods
}
