% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{probs_link}
\alias{probs_link}
\title{Takes a vector of numbers on the real line, and maps them to a unique
vector of probabilities that sum to one}
\usage{
probs_link(x)
}
\arguments{
\item{x}{The vector of values on the real line}
}
\value{
A vector of probabilities that sum to one
}
\description{
Takes a vector of numbers on the real line, and maps them to a unique
vector of probabilities that sum to one
}
