% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{profile_ll}
\alias{profile_ll}
\title{Profile (concentrated) likelihood evaluation for GAI}
\usage{
profile_ll(
  par,
  obs,
  skeleton,
  a_choice = "mixture",
  dist_choice = "P",
  spline_specs = list(df = 10, degree = 3),
  N = NULL,
  returnN = F,
  DMs = list(),
  returnA = F
)
}
\arguments{
\item{par}{A vector of estimated parameters for the model}

\item{obs}{A matrix of observations. Rows for sites and columns for
occasions.}

\item{skeleton}{The skeleton which tells this function how to \code{relist}
the estimated parameters.}

\item{a_choice}{The character name of the method used to estimate the
seasonal trend. Either "mixture", "stopover" or "splines".}

\item{dist_choice}{The distribution for the observed counts, can be "P"
(poisson), "NB" (negative binomial), "ZIP" (zero inflated poisson)}

\item{spline_specs}{A list containing the df (degrees of freedom) and the
'degree' for the splines.}

\item{N}{optionally, instead of using the profile likelihood approach to
estimate N (the superpopulation at each site), it can be supplied as a vector
or integer.}

\item{returnN}{if TRUE, returns the vector estimate of N and the matrix
a_func instead of the LL}

\item{DMs}{A list of design matrices if covariates have been included}

\item{returnA}{If TRUE, will also include the flight pattern densities}
}
\description{
A function which evaluations the profile negative log likelihood of our
observed count data, given the user model specifications of a GAI model.
}
