% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{rzip}
\alias{rzip}
\title{Zero-Inflated Poisson Random Number Generation}
\usage{
rzip(n, lambda, psi)
}
\arguments{
\item{n}{The integer number of deviates to generate}

\item{lambda}{The vector of means for each observation. Must have the same 
length n, or be of length 1 for constant means}

\item{psi}{The probability of being from the poisson distribution as
opposed to being an 'inflated' 0}
}
\value{
A vector of n points from a zero inflated poisson distribution
}
\description{
Simulates deviates from a zero-inflated poisson model given the model's
parameter values.
}
