% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{sim_n_mixture}
\alias{sim_n_mixture}
\title{Normal mixture sampling}
\usage{
sim_n_mixture(n, par, skeleton = NULL)
}
\arguments{
\item{n}{The number of samples to generate}

\item{par}{A list of parameter values. The first entry contains the}

\item{skeleton}{An object which helps to relist the vector of parameters into
the expected format.means of each component, the second the SDs, and the last
contains the weights. An exponential link is applied to the
SDs, and the probs_link function is used on the weights.
probs_link is simply a variant of the qlogis link}
}
\value{
A vector of n generated normal-mixture samples.
}
\description{
produces samples from a mixture of n normal distributions.
}
