% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{simulate_data}
\alias{simulate_data}
\title{Simulate data from the GAI}
\usage{
simulate_data(
  par,
  nT = 25,
  totals = rep(500, 50),
  a_type = "mixture",
  dist_type = "P",
  options = list(),
  thin = 0.4,
  returnDF = F
)
}
\arguments{
\item{par}{The vector of parameter values to be used}

\item{nT}{The integer number of sampling occasions}

\item{totals}{The integer site total for each site}

\item{a_type}{The character name of the seasonal distribution}

\item{dist_type}{The character name of the count distribution}

\item{options}{A list of options to specify behaviours of the seasonal
distribution. Inlcludes B, the number of broods. Currently covariates are
not supported.}

\item{thin}{The proportion of sampling opportunities which should be missed,
and replaced with an NA.}

\item{returnDF}{If TRUE, returns a data.frame with site, occasion, and count
columns, rather than a matrix of counts}
}
\value{
A matrix of simulated counts for each site (rows) and each occasion
(columns).
}
\description{
Produces a simulated data set for observed counts of individuals at each of
nS sites for nT occasions
}
