% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sum_to_one}
\alias{sum_to_one}
\title{Rescale a numeric vector so that its entries sum to one}
\usage{
sum_to_one(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\description{
Rescale a numeric vector so that its entries sum to one
}
