% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{summary.GAI}
\alias{summary.GAI}
\title{GAI model summary}
\usage{
\method{summary}{GAI}(GAIobj)
}
\arguments{
\item{GAIobj}{An object produced by using fit_GAI for model fitting}
}
\value{
An object object of class "summary.GAI"
}
\description{
Produces a set of model summaries for fitted GAI models. Includes the AIC, 
the MLEs and the average estmated site total across sites.
}
