% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{test_dependencies}
\alias{test_dependencies}
\title{Testing if dependencies are correctly loaded}
\usage{
test_dependencies()
}
\description{
As a simple check to verify that the GAI package has succesfully been able
to load its dependencies.
}
