% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{thin_NA}
\alias{thin_NA}
\title{Introduce NAs into matrices}
\usage{
thin_NA(matvec, alpha = 0.3)
}
\arguments{
\item{matvec}{The input matric or vector to be thinned}

\item{alpha}{The probability of a point being replaced}
}
\value{
The object matvec, but with some elements replaced by NA values
}
\description{
Thin a matrix of count data by replacing each of its values by an NA
observation with a fixed probability
}
