% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{transform_bootstrap_parameters}
\alias{transform_bootstrap_parameters}
\title{Transforms bootstrap confidence intervals for custom covariate values.}
\usage{
transform_bootstrap_parameters(
  bootstrap_param_output,
  GAI_fit,
  covariate_data_frame
)
}
\arguments{
\item{bootstrap_param_output}{A table of parameter estimates obtained by
using rGAI's bootstrap function with transform = F. Custom values produced
manually can also be passed.}

\item{GAI_fit}{An object produced by using fit_GAI for model fitting}

\item{covariate_data_frame}{A data.frame containing the range of covariate
values for which parameter scale transformations are desired.}
}
\description{
Transforms bootstrap confidence intervals of parameter values to the
parameter scale for a range of custom covariate values. Typically used to 
provide information to produce plots of parameter estimates against 
covariate value, with confidence intervals
}
