% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{transform_starting_values}
\alias{transform_starting_values}
\title{Transform starting values to the link scale}
\usage{
transform_starting_values(
  starting_values,
  a_choice,
  dist_choice,
  options,
  DF = list()
)
}
\arguments{
\item{starting_values}{A named list including starting values for "mu",
"sigma", "w" and "dist.par" (in the case of a non-poisson model)}

\item{a_choice}{The choice of flight path distribution}

\item{dist_choice}{The choice of observed count distribution}

\item{options}{The list of model options that specify the number of broods, 
spline parameters, covariate formulas, etc.}

\item{DF}{A data.frame which must be included when covariate formulas are
specified, so that the correct number of parameters required by each 
formula can be determined. Calling extract_counts on this data.frame first
to sanitize it is recommended}
}
\value{
A named list of starting parameter values, all on the link scale.
}
\description{
Takes in starting guesses for parameter values, and maps them to the real 
valued link scale. Even parameters already on the real scale (such as
means) are transformed, to ensure they are uniquely determined. Starting
values for splines and covariates are always set to 0, as are starting
values for parameters that aren't specified.
}
