% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{fit_GAI}
\alias{fit_GAI}
\title{GAI model fitting}
\usage{
fit_GAI(
  start,
  DF,
  a_choice = "mixture",
  dist_choice = "P",
  options = list(),
  tol = 0.001,
  maxiter = 1000,
  verbose = F,
  hessian = F,
  bootstrap = NULL,
  method = "Nelder-Mead"
)
}
\arguments{
\item{start}{The vector of start guesses for the parameters.}

\item{DF}{The data.frame of observations. Should contain columns 'site', 
'occasion' and 'count'. occasion and count should be integers, and missing 
entries in the count column are allowed to take the value NA. Any covariate
which is referred to in a formula specified in the options argument should
be present as a column in DF.}

\item{a_choice}{Choice of flight path function, which describes the seasonal
variation in counts over the monitoring period. Possible options are
"mixture": A normal mixture, with as many components as desired, will be used
to model the seasonal variation in counts.
"stopover": The Matechou et al. (2014) stopver model for seasonal variation 
will be used.
"splines": A general additive model, using splines, will model the seasonal 
variation. The degree and degrees of freedom of the splines can be specified
in the 'options' argument.}

\item{dist_choice}{The distribution for observed counts of individuals.
"P" indicates a poisson distribution, "NB" a, negative binomial, and "ZIP", 
the zero-inflated poisson distribution.}

\item{options}{A list containing different specifications, which vary
depending on the model. For stopover and mixture models this contains B (the
number of broods), shared_sigma (boolean denoting if the SDs are the same for
each component), mu_formula (specifying a formula which describes a covariate
dependency for the mean arrival times for each brood), and sd_formula
(similar, for the SD of each brood). When splines are being used, options 
'degree' and 'DF' can be included to specify the degree of the polynomial used
by the splines, and the total degrees of freedom of the seasonal flight path
model, respectively.}

\item{tol}{The tolerance for the NB and ZIP iterative solvers. The solver
stops when the difference between the negative log likelihood from one
iteration to the next falls beneath this threshold.}

\item{maxiter}{The maximum number of iterations to perform when iterating for
the NB or ZIP likelihoods}

\item{verbose}{If TRUE, the function will print various updates while it
iterates for NB and for ZIP models.}

\item{hessian}{if TRUE, refits the model one last time with hessian = TRUE in
optim, usually so that this can be used to conduct a bootstrap}

\item{bootstrap}{Always NULL for the user.. Internally, bootstraps use
this argument to pass in sanitised precalculations to speed up the refitting
process of 'refit-the-model' bootstraps}

\item{method}{A character for the method which should be used to find MLEs.
See optim documentation for options and further detail. SANN (simulated 
annealing) can be a good method of fine-tuning starting points, if one is 
afraid their MLEs are caught in a local maxima.}
}
\value{
a fitted optim object
}
\description{
Fits the GAI model to a set of data, with a choice of normal, stopover or
spline fits for the seasonal arrival times, as well as poisson, zero-inflated
poisson, or negative binomially distribted count observations.
}
