% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{means_link}
\alias{means_link}
\title{Takes a vector of numbers on the real line, and maps them to a unique
vector of means}
\usage{
means_link(means)
}
\arguments{
\item{x}{The vector of values on the real line, which tells us how far
the next value is from the previous mean}
}
\value{
A vector of means
}
\description{
Takes a vector of numbers on the real line, and maps them to a unique
vector of means
}
