% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{probs_link}
\alias{probs_link}
\title{Probability link function}
\usage{
probs_link(p)
}
\arguments{
\item{p}{A vector of real valued numerics}
}
\value{
A vector weights that sum to one, distributed according to p
}
\description{
Transforms a vector of numbers on the real line to a vector of weights that
sum to one
}
\examples{
probs_link(c(-1, 0, 1))
}
