% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{sum_to_one}
\alias{sum_to_one}
\title{Normalisation function}
\usage{
sum_to_one(v)
}
\arguments{
\item{v}{A vector of real valued parameters}
}
\value{
A vector of scaled values that add up to 1
}
\description{
Scales an input vector to have a sum of 1
}
